pyproject.toml
[project]
name = "ucp-conformance-tests"
version = "0.1.0"
description = "Test the conformance of a merchant server implementation with the UCP spec"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "UCP Team" }
]
dependencies = [
    "absl-py>=1.0.0",
    "httpx>=0.26.0",
    "pydantic>=2.12.0",
    "ucp-sdk",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.38.0",
]

[dependency-groups]
dev = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv.sources]
# The relative path is stored here
ucp-sdk = { path = "../sdk/python/", editable = true }
